# Optimize android notification saving