--- a/com.unity.mobile.notifications/Runtime/Android/Plugins/com/unity/androidnotifications/UnityNotificationRestartOnBootReceiver.java
+++ b/com.unity.mobile.notifications/Runtime/Android/Plugins/com/unity/androidnotifications/UnityNotificationRestartOnBootReceiver.java
@@ -49,7 +49,9 @@ public class UnityNotificationRestartOnBootReceiver extends BroadcastReceiver {
                     PendingIntent broadcast = UnityNotificationManager.getBroadcastPendingIntent(context, id, intent, PendingIntent.FLAG_UPDATE_CURRENT);
                     UnityNotificationManager.scheduleNotificationIntentAlarm(context, repeatInterval, fireTime, broadcast);
                 } else {
-                    UnityNotificationManager.deleteExpiredNotificationIntent(context, Integer.toString(id));
+                    String idStr = String.valueOf(id);
+                    UnityNotificationManager.removeScheduledNotificationID(context, idStr);
+                    UnityNotificationManager.deleteExpiredNotificationIntent(context, idStr);
                 }
             }
         }