Optimizes notification scheduling and cancelling by optimizing/multithreading saving of data. The issues:
https://github.com/Unity-Technologies/com.unity.mobile.notifications/issues/113
https://github.com/Unity-Technologies/com.unity.mobile.notifications/issues/112

The key changes are to do less redundant stuff and to offload some of the work to other threads. For user there should be no behavior changes.
Some numbers (Nokia 7 Plus):
- Scheduling 100 notifications from ~1500ms when down to ~500ms
- Cancelling all 100 notifications from ~130ms when down to ~1ms (almost all work moved to other thread)

Sending notifications could be improved further, but the best approach for that is to performance-wise improve the core AndroidJNI first.