@aurimasc 2021-11-15
> Is there an ETA for when this fix will be released?

We are wrapping up for the next release. There is one important bug for iOS we need to get down first. Once that is down, it should a matter of week or so get new version out plus some more time to get the new version verified. The concerning factor is that people tent to go on vacation around Christmas time, so there is a bit of time pressure to do it before that.